requirements.txt
asgiref==3.8.1
azure-cosmos==4.7.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2

django==5.2
django-cors-headers==4.4.0
dnspython==2.6.1
idna==3.10
PyJWT==2.9.0
pymongo==4.8.0
pillow==10.2.0
pytz==2025.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
numpy==2.2.5

imageio==2.37.0 
imageio_ffmpeg==0.6.0 
proglog==0.1.11 
tqdm==4.67.1
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86

decorator==4.4.2 
moviepy==1.0.3
openai==1.88.0

xhtml2pdf==0.2.15
reportlab==4.0.9
rembg==2.0.66
qrcode==7.4.2

azure-core==1.34.0 
azure-storage-blob==12.25.1 
isodate==0.7.2

pycryptodome==3.23.0
azure-identity==1.23.1 
azure-keyvault-keys==4.11.0 
msal==1.32.3 
msal-extensions==1.3.1